Rename core/framework/math_utils.h to core/framework/math.h.
