+++ onnxruntime/core/framework/math.h
