+++ onnxruntime/core/providers/cpu/math/element_wise_ops.cc
@@ -4,7 +4,7 @@
 #include "core/providers/cpu/math/element_wise_ops.h"
 
 #include "core/framework/data_types_internal.h"
-#include "core/framework/math_utils.h"
+#include "core/framework/math.h"
 #include "core/providers/cpu/tensor/utils.h"
 #include "core/providers/op_kernel_type_control.h"
 #include <unsupported/Eigen/SpecialFunctions>
