+++ onnxruntime/core/providers/cpu/tensor/isnan.cc
@@ -4,7 +4,7 @@
 #include "core/providers/cpu/tensor/isnan.h"
 
 #include "core/common/common.h"
-#include "core/framework/math_utils.h"
+#include "core/framework/math.h"
 #include "core/framework/tensor.h"
 #include "Eigen/src/Core/arch/Default/Half.h"
