+++ orttraining/orttraining/training_ops/cpu/nn/dropout_7.cc
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 #include "orttraining/training_ops/cpu/nn/dropout_7.h"
-#include "core/framework/math_utils.h"
+#include "core/framework/math.h"
 
 namespace onnxruntime {
